refactor: extract ClaudeLoginCoordinator to auth_coordinator.py
Extract authentication coordination logic from claude_service.py into a
dedicated auth_coordinator.py module. This improves separation of concerns
and makes the codebase easier to maintain.

Changes:
- Create backend/src/palette_sidecar/auth_coordinator.py
- Move ClaudeAuthStatus, ClaudeLoginCoordinator, and auth functions
- Update imports in claude_service.py to use new module
- Update test imports to use auth_coordinator

Impact:
- claude_service.py: 872 → 593 lines (-279 lines, -32%)
- All 6 backend tests pass
- Zero behavioral changes
